[AIRFLOW-5775] Migrate AWS Batch components to /providers/aws [AIP-21]
